Reconstruct Sketch file
I don't know what happened, but the Sketch file with the latest versions
of all the illustrations was gone, and I couldn't find it anywhere.
There were only a much earlier version of the Sketch file in the repo.

However, thanks to the illustrations themselves all being present in SVG
format, I could reconstruct (approsximately) the latest version of the
Sketch file.